Added types and cleaned up schema location
